Break `modules.rb` up into files for real names
It's a little surprising to require a file and then get top-level types
loaded with a different name. If I load a thing called `modules`, I 
expect a type named `Modules` to be loaded, not `Toolbox` and `Pusheen`

So there are two options to fix it:

1. Change `modules.rb` to have a top-level module with `Toolbox` and
`Pusheen` underneath it:

```
module Modules
  module Toolbox
    ...
  end
end
```

Then you'd still `require 'modules'` and include `Modules::Toolbox` 

2. Pull both modules out into their own files.

I did #2 because the two modules have nothing to do with one another,
and naming a module `Modules` is confusing, since it's so close to
`Module`, which is a keyword
