+++ lib/search.rb
@@ -1,4 +1,4 @@
-require '../lib/modules'
+require '../lib/toolbox'
 require 'json'
 require 'pry'
